refactor(ai-proxy): rename routes.ts to types.ts
Better reflects that this file contains type definitions, not routing logic.
